+++ apps/ui/components/search/SeasonalSearchForm.tsx
@@ -14,7 +14,7 @@ import { participantCountOptions } from "@/modules/const";
 import { MediumButton } from "@/styles/util";
 import { JustForDesktop, JustForMobile } from "@/modules/style/layout";
 import { useSearchModify, useSearchValues } from "@/hooks/useSearchValues";
-import { FilterTagList } from "../single-search/FilterTagList";
+import { FilterTagList } from "./FilterTagList";
 import { ParsedUrlQuery } from "node:querystring";
 import { ControlledSelect } from "@/components/common/ControlledSelect";
 import { ControlledMultiSelect } from "./ControlledMultiSelect";
