+++ apps/ui/pages/search/single.tsx
@@ -19,10 +19,10 @@ import { Container } from "common";
 import { filterNonNullable } from "common/src/helpers";
 import { isBrowser } from "@/modules/const";
 import { RESERVATION_UNITS } from "@/modules/queries/reservationUnit";
-import { SingleSearchForm } from "@/components/single-search/SearchForm";
+import { SingleSearchForm } from "@/components/search/SingleSearchForm";
 import Sorting from "@/components/form/Sorting";
 import ListWithPagination from "@/components/common/ListWithPagination";
-import ReservationUnitCard from "@/components/single-search/ReservationUnitCard";
+import ReservationUnitCard from "@/components/search/SingleSearchReservationUnitCard";
 import BreadcrumbWrapper from "@/components/common/BreadcrumbWrapper";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
 import { createApolloClient } from "@/modules/apolloClient";
